+++ packages/aws-amplify-angular/index.ts
@@ -1,32 +1,43 @@
 export { AmplifyAngularModule } from './src/aws-amplify-angular.module';
-// export { AmplifyIonicModule } from './src/aws-amplify-ionic-module';
-
-
+export { AmplifyIonicModule } from './src/aws-amplify-ionic-module';
 export { AmplifyService}  from './src/providers/amplify.service';
 
-export { AuthenticatorComponent } from './src/components/authenticator/authenticator/authenticator.factory';
-export { AuthenticatorComponentCore } from './src/components/authenticator/authenticator/authenticator.component.core';
-export { ConfirmSignInComponent } from './src/components/authenticator/confirm-sign-in-component/confirm-sign-in.factory';
-export { ConfirmSignInComponentCore } from './src/components/authenticator/confirm-sign-in-component/confirm-sign-in-component.core';
-export { ConfirmSignUpComponent } from './src/components/authenticator/confirm-sign-up-component/confirm-sign-up.factory';
-export { ConfirmSignUpComponentCore } from './src/components/authenticator/confirm-sign-up-component/confirm-sign-up.component.core';
-export { SignInComponent } from './src/components/authenticator/sign-in-component/sign-in.component.factory';
-export { SignInComponentCore } from './src/components/authenticator/sign-in-component/sign-in.component.core';
-export { SignUpComponent } from './src/components/authenticator/sign-up-component/sign-up.factory';
-export { SignUpComponentCore } from './src/components/authenticator/sign-up-component/sign-up.component.core';
-export { RequireNewPasswordComponent } from './src/components/authenticator/require-new-password-component/require-new-password.factory';
-export { RequireNewPasswordComponentCore } from './src/components/authenticator/require-new-password-component/require-new-password.component.core';
-export { GreetingComponent } from './src/components/authenticator/greeting-component/greeting.factory';
-export { GreetingComponentCore } from './src/components/authenticator/greeting-component/greeting.component.core';
-export { ForgotPasswordComponent } from './src/components/authenticator/forgot-password-component/forgot-password.factory';
-export { ForgotPasswordComponentCore } from './src/components/authenticator/forgot-password-component/forgot-password.component.core';
-export { S3AlbumComponent } from './src/components/storage/s3-album-component/s3-album.factory';
-export { S3AlbumComponentCore } from './src/components/storage/s3-album-component/s3-album.component.core';
-export { S3ImageComponent } from './src/components/storage/s3-image-component/s3-image.factory';
-export { S3ImageComponentCore } from './src/components/storage/s3-image-component/s3-image.component.core';
-export { PhotoPickerComponent } from './src/components/storage/photo-picker-component/photo-picker.factory';
-export { PhotoPickerComponentCore } from './src/components/storage/photo-picker-component/photo-picker.component.core';
-export { ChatBotComponent } from './src/components/interactions/chatbot/chatbot.factory';
-export { ChatbotComponentCore } from './src/components/interactions/chatbot/chatbot.component.core'; 
-export { DynamicComponentDirective } from './src/directives/dynamic.component.directive';
-export { FormComponent } from './src/components/common/form.component';
+// export { AuthenticatorComponent } from './src/components/authenticator/authenticator/authenticator.factory';
+// export { AuthenticatorComponentCore } from './src/components/authenticator/authenticator/authenticator.component.core';
+// export { ConfirmSignInComponent } from './src/components/authenticator/confirm-sign-in-component/confirm-sign-in.factory';
+// export { ConfirmSignInComponentCore } from './src/components/authenticator/confirm-sign-in-component/confirm-sign-in-component.core';
+// export { ConfirmSignUpComponent } from './src/components/authenticator/confirm-sign-up-component/confirm-sign-up.factory';
+// export { ConfirmSignUpComponentCore } from './src/components/authenticator/confirm-sign-up-component/confirm-sign-up.component.core';
+// export { SignInComponent } from './src/components/authenticator/sign-in-component/sign-in.component.factory';
+// export { SignInComponentCore } from './src/components/authenticator/sign-in-component/sign-in.component.core';
+// export { SignUpComponent } from './src/components/authenticator/sign-up-component/sign-up.factory';
+// export { SignUpComponentCore } from './src/components/authenticator/sign-up-component/sign-up.component.core';
+// export { RequireNewPasswordComponent } from './src/components/authenticator/require-new-password-component/require-new-password.factory';
+// export { RequireNewPasswordComponentCore } from './src/components/authenticator/require-new-password-component/require-new-password.component.core';
+// export { GreetingComponent } from './src/components/authenticator/greeting-component/greeting.factory';
+// export { GreetingComponentCore } from './src/components/authenticator/greeting-component/greeting.component.core';
+// export { ForgotPasswordComponent } from './src/components/authenticator/forgot-password-component/forgot-password.factory';
+// export { ForgotPasswordComponentCore } from './src/components/authenticator/forgot-password-component/forgot-password.component.core';
+// export { S3AlbumComponent } from './src/components/storage/s3-album-component/s3-album.factory';
+// export { S3AlbumComponentCore } from './src/components/storage/s3-album-component/s3-album.component.core';
+// export { S3ImageComponent } from './src/components/storage/s3-image-component/s3-image.factory';
+// export { S3ImageComponentCore } from './src/components/storage/s3-image-component/s3-image.component.core';
+// export { PhotoPickerComponent } from './src/components/storage/photo-picker-component/photo-picker.factory';
+// export { PhotoPickerComponentCore } from './src/components/storage/photo-picker-component/photo-picker.component.core';
+// export { ChatBotComponent } from './src/components/interactions/chatbot/chatbot.factory';
+// export { ChatbotComponentCore } from './src/components/interactions/chatbot/chatbot.component.core'; 
+// export { DynamicComponentDirective } from './src/directives/dynamic.component.directive';
+// export { FormComponent } from './src/components/common/form.component';
+
+// export { AuthenticatorIonicComponent } from './components/authenticator/authenticator/authenticator.component.ionic';
+// export { ConfirmSignInComponentIonic } from './components/authenticator/confirm-sign-in-component/confirm-sign-in-component.ionic';
+// export { ConfirmSignUpComponentIonic } from './components/authenticator/confirm-sign-up-component/confirm-sign-up.component.ionic';
+// export { ForgotPasswordComponentIonic } from './components/authenticator/forgot-password-component/forgot-password.component.ionic';
+// export { GreetingComponentIonic } from './components/authenticator/greeting-component/greeting.component.ionic';
+// export { SignInComponentIonic } from './components/authenticator/sign-in-component/sign-in.component.ionic';
+// export { SignUpComponentIonic } from './components/authenticator/sign-up-component/sign-up.component.ionic';
+// export { RequireNewPasswordComponentIonic } from './components/authenticator/require-new-password-component/require-new-password.component.ionic';
+// export { PhotoPickerIonicComponent } from './components/storage/photo-picker-component/photo-picker.component.ionic';
+// export { S3AlbumComponentIonic } from './components/storage/s3-album-component/s3-album.component.ionic';
+// export { S3ImageComponentIonic } from './components/storage/s3-image-component/s3-image.component.ionic';
+// export { ChatbotComponentIonic } from './components/interactions/chatbot/chatbot.component.ionic';
