+++ packages/aws-amplify-angular/src/components/authenticator/forgot-password-component/forgot-password.component.ionic.ts
@@ -1,8 +1,9 @@
 import { Component, Input } from '@angular/core';
-import { AmplifyService, AuthState } from '../providers';
-import { RequireNewPasswordComponentCore } from '../components/authenticator/require-new-password-component/require-new-password.component.core';
+import { AmplifyService, AuthState } from '../../../providers';
+import { ForgotPasswordComponentCore } from './forgot-password.component.core';
+// import { includes } from '../common';
 
-const templatetemp = `
+const template = `
 <div class="amplify-form-container" *ngIf="_show">
   <div class="amplify-form-body">
 
@@ -63,47 +64,11 @@ const templatetemp = `
 
 `
 
-const template = `
-<div class="amplify-form-container" *ngIf="_show">
-  <div class="amplify-form-body">
-  <div class="amplify-form-row">
-      <div class="amplify-form-cell-left">
-        <a class="amplify-form-link"
-          (click)="onSignIn()"
-        >Back to Sign In</a>
-      </div>
-    </div>
-    <ion-list>
-      <ion-item>
-    
-        <ion-label stacked>Password</ion-label>
-        <ion-input 
-        #password
-        type="password" 
-        (keyup)="setPassword(password.value)"
-        (keyup.enter)="onSubmit()"
-        ></ion-input>
-      </ion-item>
-    </ion-list>
-    
-    <div class="amplify-form-row">
-      <ion-button
-        (click)="onSubmit()"
-      >Submit</ion-button>
-    </div>
-  </div>
-  <div class="amplify-form-footer">
-    <div class="amplify-form-message-error" *ngIf="errorMessage">{{ errorMessage }}</div>
-  </div>
-  
-</div>
-`
-
 @Component({
-  selector: 'amplify-auth-require-new-password-ionic',
+  selector: 'amplify-auth-forgot-password-ionic',
   template: template
 })
-export class RequireNewPasswordComponentIonic extends RequireNewPasswordComponentCore {
+export class ForgotPasswordComponentIonic extends ForgotPasswordComponentCore {
 
 
   constructor(amplifyService: AmplifyService) {
