+++ packages/aws-amplify-angular/src/components/authenticator/greeting-component/greeting.component.ionic.ts
@@ -1,7 +1,6 @@
 import { Component, Input } from '@angular/core';
-import { AmplifyService, AuthState } from '../providers';
-import { GreetingComponentCore } from '../components/authenticator/greeting-component/greeting.component.core';
-// import { includes } from '../common';
+import { AmplifyService, AuthState } from '../../../providers';
+import { GreetingComponentCore } from './greeting.component.core';
 
 const template = `
 <div class="amplify-greeting">
