+++ packages/aws-amplify-angular/src/components/authenticator/sign-up-component/sign-up.component.ionic.ts
@@ -1,21 +1,19 @@
 import { Component, Input } from '@angular/core';
-import { AmplifyService, AuthState } from '../providers';
-import { SignInComponentCore } from '../components/authenticator/sign-in-component/sign-in.component.core';
-// import { includes } from '../common';
+import { AmplifyService, AuthState } from '../../../providers';
+import { SignUpComponentCore } from './sign-up.component.core';
 
 const template = `
 <div class="amplify-form-container" *ngIf="_show">
-  <div class="amplify-form-header"></div>
   <div class="amplify-form-body">
-    
+
     <div class="amplify-form-row">
-      
+
       <div class="amplify-form-cell-left">
         <a class="amplify-form-link"
           (click)="onSignIn()"
         >Sign In</a>
       </div>
-      
+
       <div class="amplify-form-cell-right">
         <a class="amplify-form-link"
           (click)="onSignUp()"
@@ -24,13 +22,15 @@ const template = `
 
     </div>
 
+
     <ion-list>
 
       <ion-item>
         <ion-label stacked>Username</ion-label>
-        <ion-input type="text" 
+        <ion-input 
+          #username
+          type="text" 
           (keyup)="setUsername($event.target.value)"
-          [value]="username"
         ></ion-input>
       </ion-item>
     
@@ -40,44 +40,55 @@ const template = `
           #password
           type="password" 
           (keyup)="setPassword(password.value)"
-          (keyup.enter)="onSignIn()"
+          (keyup.enter)="onSignUp()"
+        ></ion-input>
+      </ion-item>
+
+      <ion-item>
+        <ion-label stacked>Email</ion-label>
+        <ion-input 
+          #email
+          type="text" 
+          (keyup)="setEmail(email.value)"
+        ></ion-input>
+      </ion-item>
+
+      <ion-item>
+        <ion-label stacked>Phone Number</ion-label>
+        <ion-input 
+          #phone_number
+          type="text" 
+          (keyup)="setPhoneNumber(phone_number.value)"
         ></ion-input>
       </ion-item>
     
     </ion-list>
 
     <div class="amplify-form-row">
-
       <div class="amplify-form-cell-right">
         <a class="amplify-form-link"
-          (click)="onForgotPassword()"
-        >Forgot Password</a>
+          (click)="onConfirmSignUp()"
+        >Confirm a Code</a>
       </div>
-
     </div>
-
-    <ion-button
-      (click)="onSignIn()"
-    >Sign In</ion-button>
-
+    <ion-button 
+      (click)="onSignUp()"
+    >Sign Up</ion-button>
   </div>
-
   <div class="amplify-form-footer">
     <div class="amplify-form-message-error" *ngIf="errorMessage">{{ errorMessage }}</div>
   </div>
-
 </div>
-
 `
 
 @Component({
-  selector: 'amplify-auth-sign-in-ionic',
+  selector: 'amplify-auth-sign-up-ionic',
   template: template
 })
-export class SignInComponentIonic extends SignInComponentCore {
+export class SignUpComponentIonic extends SignUpComponentCore {
 
   constructor(amplifyService: AmplifyService) {
-    super(amplifyService);    
+    super(amplifyService)
   }
-  
+
 }
