+++ packages/aws-amplify-angular/tsconfig-angular-ionic.json
